backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-multipart==0.0.20
pillow==11.1.0
